+++ NEWS.md
@@ -1,3 +1,5 @@
+* Fixed a bug in test session resuming using `get_p_id()` pages.
+
 # psychTestR 2.2.0
 
 * Allowing extra attributes to be passed to `trigger_button()`.